añadiendo manejo de estado suspendido y reactivacion para actividad y tarea-actividad

+++ routes/api/activity.js
@@ -137,7 +137,10 @@ router.post('/delete', [
         if(!activity){
             res.status(404).json({errors: [{msg: "La actividad a eliminar no existe."}]});
         };
-
+        //validación de que solo se puedan eliminar actividades creadas
+        if(!(activity.status=="CREADA")){
+            return res.status(404).json({errors: [{msg: "La Actividad se encuentra asociada a tareas activas"}]});
+        }
         //eliminación de la actividad
         await Activity.findOneAndRemove({_id: id});
         
@@ -152,6 +155,118 @@ router.post('/delete', [
 
         res.json({msg: 'Etapa eliminada'});
         
+    } catch (err) {
+        console.error(err.message);
+        res.status(500).send('Server Error: ' + err.message);
+    }
+
+});
+
+
+// @route POST api/activity/suspense
+// @desc  suspende una actividad segun id
+// @access Public
+router.post('/suspense', [
+    check('id', 'Id es requerido').not().isEmpty(),
+    check('idUserCreate', 'El Usuario no está autenticado').not().isEmpty(),
+    check('reason',"La razón es necesario").not().isEmpty(),
+], async(req, res) => {
+
+    const errors = validationResult(req);
+    if(!errors.isEmpty()){
+        return res.status(400).json({ errors: errors.array() });
+    }
+
+    const id = req.body.id;
+    const idUserCreate = req.body.idUserCreate;
+    const reason = req.body.reason;
+    
+    try {
+
+        let activity = await Activity.findById(id);
+        if(!activity){
+            return res.status(404).json({errors: [{msg: "La Actividad no existe."}]});
+        }else{ //actividad existente.           
+            if(!(activity.status=="ACTIVA")){
+                return res.status(404).json({errors: [{msg: "La Actividad no se encuentra activa para suspenderla"}]});
+            }
+            // iterar por tareas asignadas y a las "ACTIVA", cambiar por "SUSPENDIDA"
+            //
+            //
+            //-------------FALTA!!!
+            //
+            //
+            //Cambiar estado de la actividad  a "SUSPENDIDA" y generar historial. Agendar "quien" lo suspende
+            
+            let posLastHistoryActivity = activity.history.length - 1;        
+        
+            let idLastHistoryActivity = activity.history[posLastHistoryActivity]._id
+
+            let dateToday = Date.now();  
+
+            let reasonAdd = "-";
+            if (reason !== ""){
+                reasonAdd = reason;
+            };
+
+            await Activity.findOneAndUpdate({_id: id,"history._id":idLastHistoryActivity}, {$set:{"history.$.dateDown":dateToday}});
+            
+            await Activity.findOneAndUpdate({_id: id}, {$set:{status:"SUSPENDIDA"},$push: { history: {status:"SUSPENDIDA",dateUp:dateToday,reason:reasonAdd,idUserChanged:idUserCreate}}});
+
+            res.json({msg: 'Actividad suspendida'});
+        }
+    } catch (err) {
+        console.error(err.message);
+        res.status(500).send('Server Error: ' + err.message);
+    }
+
+});
+
+// @route POST api/activity/reactivate
+// @desc  REACTIVA una actividad segun id
+// @access Public
+router.post('/reactivate', [
+    check('id', 'Id es requerido').not().isEmpty(),
+    check('idUserCreate', 'El Usuario no está autenticado').not().isEmpty(),
+], async(req, res) => {
+
+    const errors = validationResult(req);
+    if(!errors.isEmpty()){
+        return res.status(400).json({ errors: errors.array() });
+    }
+
+    const id = req.body.id;
+    const idUserCreate = req.body.idUserCreate;
+    
+    try {
+
+        let activity = await Activity.findById(id);
+        if(!activity){
+            return res.status(404).json({errors: [{msg: "La Actividad no existe."}]});
+        }else{ //etapa existente.
+            if(!(activity.status=="SUSPENDIDA")){
+                return res.status(404).json({errors: [{msg: "La Actividad no se encuentra suspendida para reactivarla"}]});
+            }
+            // iterar por cada tareas asignadas y a las "SUSPENDIDA", cambiar por "ACTIVA"
+            //
+            //
+            //-------------FALTA!!!
+            //
+            //
+            //Cambiar estado de la estapa a "ACTIVA" y generar historial. Agendar "quien" lo activa
+            
+            let posLastHistoryActivity = activity.history.length - 1;        
+        
+            let idLastHistoryActivity = activity.history[posLastHistoryActivity]._id
+
+            let dateToday = Date.now();  
+
+            await Activity.findOneAndUpdate({_id: id,"history._id":idLastHistoryActivity}, {$set:{"history.$.dateDown":dateToday}});
+            
+            await Activity.findOneAndUpdate({_id: id}, {$set:{status:"ACTIVA"},$push: { history: {status:"ACTIVA",dateUp:dateToday,idUserChanged:idUserCreate}}});
+
+            res.json({msg: 'Actividad activada'});
+        }
     } catch (err) {
         console.error(err.message);
         res.status(500).send('Server Error: ' + err.message);
